refactor: ubah route untul apl1 dan apl2 menjadi /api/assessment/apl...
