+++ src/app.ts
@@ -18,22 +18,24 @@ import userRoutes from './modules/user/user.routes';
 import authRoutes from './modules/auth/auth.routes';
 import dashboardRoutes from './modules/admin/dashboard/dashboard.routes';
 import assesseeRoutes from './modules/assessee/assessee.routes';
-import unitCompetencyRoutes from './modules/unit-competency/unit-competency.routes';
 import occupationRoutes from './modules/occupation/occupation.routes';
-import apl1Routes from './modules/apl1/apl1.routes';
-import apl2Routes from './modules/apl2/apl2.routes';
+
+// DONE - DONE
+import apl1Routes from './modules/assessement/apl1/apl1.routes';
+import apl2Routes from './modules/assessement/apl2/apl2.routes';
+
+app.use('/api/assessment/apl1', apl1Routes);
+app.use('/api/assessment/apl2', apl2Routes);
+// DONE - DONE
 
 app.use('/api/users', userRoutes);
 app.use('/api/auth', authRoutes);
 app.use('/api/assessees', assesseeRoutes);
 app.use('/api/dashboard', dashboardRoutes);
-app.use('/api/unit-competencies', unitCompetencyRoutes);
 app.use('/api/occupations', occupationRoutes);
-app.use('/api/apl1', apl1Routes);
-app.use('/api/apl2', apl2Routes);
 
 /* -------- TODO --------
-  
+  TODO: apl1 url untuk gambar docs
 */
