--- src/modules/unit-competency/unit-competency.type.ts
@@ -1,14 +1,0 @@
-export interface UnitCompetencyRequest {
-    assessment_id: number;
-    unit_code: string;
-    title: string;
-}
-
-export interface UnitCompetencyResponse {
-    id: number;
-    assessment_id: number;
-    unit_code: string;
-    title: string;
-    assessment?: any;
-    elements?: any[];
-}
